Refactor sql test
