+++ extensions/tests/postgres_test.go
@@ -17,21 +17,23 @@
 // "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
-// under the License.    
+// under the License.
 
 package tests
 
 import (
 	"fmt"
+	"testing"
+	"time"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/xdblab/xdb/common/log"
 	"github.com/xdblab/xdb/config"
 	"github.com/xdblab/xdb/extensions"
 	"github.com/xdblab/xdb/extensions/postgres"
 	"github.com/xdblab/xdb/extensions/postgres/postgrestool"
+
 	"github.com/xdblab/xdb/persistence/sql"
-	"testing"
-	"time"
 )
 
 func TestPostgres(t *testing.T) {
@@ -55,8 +57,8 @@ func TestPostgres(t *testing.T) {
 
 	store, err := sql.NewSQLProcessStore(*sqlConfig, log.NewDevelopmentLogger())
 	ass.Nil(err)
-	
-	testSQLBasicExecution(ass, store)
+
+	sql.SQLBasicExecutionTest(ass, store)
 
 	_ = extensions.DropDatabase(*sqlConfig, testDBName)
 	fmt.Println("testing database deleted")
